JS-2879: history copy
